extract types + logHours refactor
